refactor(router): review follow-ups + activity-constants consolidation
Review fixes:
- LitActivityBlock only subscribes to the router when it actually has an
  `activityType` — non-activity subclasses (source buttons/list, headers,
  DynamicBtn, the cloud-editor content block, …) no longer `requestUpdate()` on
  every navigation.
- `navigate()` reuses `_runEdgeHooks('beforeChange', …)` instead of an inline
  duplicate loop.
- `openModal()` / `setActivity()` now honor activity guards (so a guarded-out
  activity, e.g. an empty upload list, can't be force-opened via
  `setModalState`/`setCurrentActivity`).
- External source header [X] uses `traverse('onClose')`, consistent with the
  camera/url close buttons.
- Documented that `EdgeContext.proposed`/`defaults()` only carry a concrete
  target for `onDone` (back/cancel/close defaults are actions, not targets).

Constants consolidation:
- Replaced `LitActivityBlock.activities.X` with `ACTIVITY_TYPES.X` at all call
  sites (UploadList, StartFrom, regular/minimal/inline solutions).
- Moved `ActivityParamsMap` to `activity-constants` and removed the
  `ActivityType`/`CustomActivities`/`RegisteredActivityType` re-export from
  `LitActivityBlock`; consumers (and the unsplash `declare module`) import from
  `activity-constants` directly. Strict `ActivityId` verified intact.

Gate green: tsc, build, specs 504, e2e 187/187, lint.
