+++ src/blocks/PluginActivityRenderer/PluginActivityRenderer.ts
@@ -3,7 +3,8 @@ import { property, state } from 'lit/decorators.js';
 import { createRef, ref } from 'lit/directives/ref.js';
 import { repeat } from 'lit/directives/repeat.js';
 import type { Owned, PluginActivityRegistration, PluginRenderDispose } from '../../abstract/managers/plugin';
-import { type ActivityType, LitActivityBlock } from '../../lit/LitActivityBlock';
+import type { ActivityType } from '../../lit/activity-constants';
+import { LitActivityBlock } from '../../lit/LitActivityBlock';
 import { LitBlock } from '../../lit/LitBlock';
 import '../Modal/Modal';
 import './uc-plugin-activity-host.css';
