+++ src/blocks/StartFrom/StartFrom.ts
@@ -1,9 +1,10 @@
 import './start-from.css';
 import { html } from 'lit';
-import { type ActivityType, LitActivityBlock } from '../../lit/LitActivityBlock';
+import { ACTIVITY_TYPES, type ActivityType } from '../../lit/activity-constants';
+import { LitActivityBlock } from '../../lit/LitActivityBlock';
 
 export class StartFrom extends LitActivityBlock {
-  public override activityType: ActivityType = LitActivityBlock.activities.START_FROM;
+  public override activityType: ActivityType = ACTIVITY_TYPES.START_FROM;
 
   public override render() {
     return html` <div class="uc-content">${this.yield('')}</div> `;
