+++ src/types/index.ts
@@ -1,5 +1,5 @@
 export type { CustomConfig, CustomConfigDefinition } from '../abstract/customConfigOptions';
 export type * from '../abstract/managers/plugin/PluginTypes';
-export type { ActivityType, CustomActivities } from '../lit/LitActivityBlock';
+export type { ActivityType, CustomActivities } from '../lit/activity-constants';
 export type * from './events';
 export type * from './exported';
